+++ pkg/analyzer/lib/src/dart/element/scope.dart
@@ -35,22 +35,21 @@ class EnclosedScope implements Scope {
 
   Scope get parent => _parent;
 
-  @Deprecated('Use lookup2() that is closer to the language specification')
   @override
-  Element lookup({@required String id, @required bool setter}) {
-    var result = lookup2(id);
-    return setter ? result.setter : result.getter;
-  }
-
-  @override
-  ScopeLookupResult lookup2(String id) {
+  ScopeLookupResult lookup(String id) {
     var getter = _getters[id];
     var setter = _setters[id];
     if (getter != null || setter != null) {
       return ScopeLookupResult(getter, setter);
     }
 
-    return _parent.lookup2(id);
+    return _parent.lookup(id);
+  }
+
+  @Deprecated('Use lookup() instead')
+  @override
+  ScopeLookupResult lookup2(String id) {
+    return lookup(id);
   }
 
   void _addGetter(Element element) {
@@ -208,15 +207,8 @@ class PrefixScope implements Scope {
     }
   }
 
-  @Deprecated('Use lookup2() that is closer to the language specification')
   @override
-  Element lookup({@required String id, @required bool setter}) {
-    var result = lookup2(id);
-    return setter ? result.setter : result.getter;
-  }
-
-  @override
-  ScopeLookupResult lookup2(String id) {
+  ScopeLookupResult lookup(String id) {
     if (_deferredLibrary != null && id == FunctionElement.LOAD_LIBRARY_NAME) {
       return ScopeLookupResult(_deferredLibrary.loadLibraryFunction, null);
     }
@@ -224,6 +216,12 @@ class PrefixScope implements Scope {
     var getter = _getters[id];
     var setter = _setters[id];
     return ScopeLookupResult(getter, setter);
+  }
+
+  @Deprecated('Use lookup() instead')
+  @override
+  ScopeLookupResult lookup2(String id) {
+    return lookup(id);
   }
 
   void _add(Element element) {
@@ -323,14 +321,14 @@ class _LibraryImportScope implements Scope {
     }.toList();
   }
 
-  @Deprecated('Use lookup2() that is closer to the language specification')
   @override
-  Element lookup({@required String id, @required bool setter}) {
-    throw UnimplementedError();
+  ScopeLookupResult lookup(String id) {
+    return _nullPrefixScope.lookup(id);
   }
 
+  @Deprecated('Use lookup() instead')
   @override
   ScopeLookupResult lookup2(String id) {
-    return _nullPrefixScope.lookup2(id);
+    return lookup(id);
   }
 }